Jormungandr: activate date and pagination lookup to synthese
